+++ .eslintrc.json
@@ -1,5 +1,9 @@
 {
   "extends": "next/core-web-vitals",
-  "plugins": ["@stylexjs"],
-  "rules": { "@stylexjs/valid-styles": "error" }
+  "plugins": [
+    "@stylexjs"
+  ],
+  "rules": {
+    "@stylexjs/valid-styles": "error"
+  }
 }
