+++ src/features/update-button/index.ts
@@ -1,3 +1,2 @@
 export { TeddyCategory } from './ui/change-teddyCategory';
 export { ChangeNavBarColor } from './ui/change-navbarCategory';
-export { ChangeDay } from './ui/changeDay';
