--- src/views/wakatime/index.ts
@@ -1,1 +1,0 @@
-export { WakaTimeHome } from './ui/wakaTime';
