--- src/widgets/SelectDay/ui/selectDay.tsx
@@ -1,11 +1,0 @@
-import { ChangeDay } from '@/features/update-button';
-
-export const SelectDay = () => {
-  return (
-    <>
-      <div>
-        <ChangeDay />
-      </div>
-    </>
-  );
-};
